Recalculate task schedule when project start changes
