+++ src/main/java/com/task_management/service/impl/ProjectServiceImpl.java
@@ -9,7 +9,9 @@ import com.task_management.exception.BadRequestException;
 import com.task_management.exception.NotFoundException;
 import com.task_management.mapper.ProjectMapper;
 import com.task_management.repository.ProjectRepository;
+import com.task_management.repository.TaskRepository;
 import com.task_management.service.ProjectService;
+import com.task_management.service.TaskScheduleCalculator;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -21,7 +23,9 @@ import java.util.UUID;
 @Service @RequiredArgsConstructor @Transactional
 public class ProjectServiceImpl implements ProjectService {
     private final ProjectRepository projects;
+    private final TaskRepository tasks;
     private final ProjectMapper mapper;
+    private final TaskScheduleCalculator scheduleCalculator;
 
     @Override
     public ProjectRes create(ProjectCreateReq req) {
@@ -48,7 +52,11 @@ public class ProjectServiceImpl implements ProjectService {
         if (req.name() != null && !p.getName().equalsIgnoreCase(req.name())
                 && projects.existsByNameIgnoreCase(req.name()))
             throw new BadRequestException("Project name already exists");
+        var originalStartDate = p.getStartDate();
         mapper.update(p, req);
+        if (req.startDate() != null && !req.startDate().equals(originalStartDate)) {
+            recalculateTaskSchedule(p);
+        }
         return mapper.toRes(projects.save(p));
     }
 
@@ -56,5 +64,17 @@ public class ProjectServiceImpl implements ProjectService {
     public void delete(UUID id) {
         if (!projects.existsById(id)) throw new NotFoundException("Project not found");
         projects.deleteById(id);
+    }
+
+    private void recalculateTaskSchedule(Project project) {
+        var projectTasks = tasks.findByProjectId(project.getId());
+        if (projectTasks.isEmpty()) {
+            return;
+        }
+        for (var task : projectTasks) {
+            task.setProject(project);
+            scheduleCalculator.applyScheduleDays(task);
+        }
+        tasks.saveAll(projectTasks);
     }
 }
